Cherry-pick changes from sc_refinement
