+++ maro/cli/k8s/aks/conf.yml
@@ -1,9 +1,9 @@
 mode: ""
-azure_subscription: 03811e6d-e6a0-4ae1-92e7-2249b8cb13be
-name: yq
-location: eastus
+azure_subscription: your_azure_subscription_id
+name: your_deployment_name
+location: your_azure_service_location
 container_registry_service_tier: Standard   # "Basic", "Standard", "Premium", see https://docs.microsoft.com/en-us/azure/container-registry/container-registry-skus for details
-hardware:
+resources:
   k8s:
     vm_size: Standard_DS2_v2  # https://docs.microsoft.com/en-us/azure/virtual-machines/sizes, https://docs.microsoft.com/en-us/azure/aks/quotas-skus-regions
     vm_count: 1  # must be at least 2 for k8s to function properly.
