+++ maro/cli/k8s/aks/template.json
@@ -96,7 +96,7 @@
     {
       "name": "[parameters('aksName')]",
       "type": "Microsoft.ContainerService/managedClusters",
-      "apiVersion": "2020-03-01",
+      "apiVersion": "2021-10-01",
       "location": "[parameters('location')]",
       "properties": {
         "dnsPrefix": "maro",
